JSON responses improved

+++ src/main/java/com/artsolo/bookswap/controllers/LibraryController.java
@@ -1,5 +1,9 @@
 package com.artsolo.bookswap.controllers;
 
+import com.artsolo.bookswap.controllers.responses.ErrorDescription;
+import com.artsolo.bookswap.controllers.responses.ErrorResponse;
+import com.artsolo.bookswap.controllers.responses.MessageResponse;
+import com.artsolo.bookswap.controllers.responses.SuccessResponse;
 import com.artsolo.bookswap.services.LibraryService;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -18,35 +22,62 @@ public class LibraryController {
     }
 
     @PostMapping("/add-book")
-    public ResponseEntity<String> addBookToLibrary(@RequestBody Map<String, String> request) {
+    public ResponseEntity<?> addBookToLibrary(@RequestBody Map<String, String> request) {
         try {
-            if(libraryService.addNewBookToUserLibrary(request)) {
-                return ResponseEntity.ok("success");
+            if (request.get("userId") != null && request.get("bookId") != null) {
+                if (libraryService.addNewBookToUserLibrary(request)) {
+                    return ResponseEntity.ok().body(new MessageResponse("Book was added to library successfully"));
+                }
+                return ResponseEntity.badRequest().body(new ErrorResponse(new ErrorDescription(
+                        HttpStatus.BAD_REQUEST.value(),
+                        "Failed to add book")
+                ));
             }
-            return ResponseEntity.ok("failed");
+            return ResponseEntity.badRequest().body(new ErrorResponse(new ErrorDescription(
+                    HttpStatus.BAD_REQUEST.value(),
+                    "Bad request")
+            ));
         } catch (Exception e) {
-            return new ResponseEntity<String>("Something want wrong", HttpStatus.INTERNAL_SERVER_ERROR);
+            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(new ErrorResponse(new ErrorDescription(
+                    HttpStatus.INTERNAL_SERVER_ERROR.value(),
+                    "Internal server error")
+            ));
         }
     }
 
     @DeleteMapping("/remove-book")
-    public ResponseEntity<String> removeBookFromLibrary(@RequestBody Map<String, String> request) {
+    public ResponseEntity<?> removeBookFromLibrary(@RequestBody Map<String, String> request) {
         try {
-            if(libraryService.removeBookFromUserLibrary(request)) {
-                return ResponseEntity.ok("success");
+            if (request.get("userId") != null && request.get("bookId") != null) {
+                if (libraryService.removeBookFromUserLibrary(request)) {
+                    return ResponseEntity.ok().body(new MessageResponse("Book was removed from library successfully"));
+                }
+                return ResponseEntity.badRequest().body(new ErrorResponse(new ErrorDescription(
+                        HttpStatus.BAD_REQUEST.value(),
+                        "Book still in the library")
+                ));
             }
-            return ResponseEntity.ok("failed");
+            return ResponseEntity.badRequest().body(new ErrorResponse(new ErrorDescription(
+                    HttpStatus.BAD_REQUEST.value(),
+                    "Bad request")
+            ));
         } catch (Exception e) {
-            return new ResponseEntity<String>("Something want wrong", HttpStatus.INTERNAL_SERVER_ERROR);
+            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(new ErrorResponse(new ErrorDescription(
+                    HttpStatus.INTERNAL_SERVER_ERROR.value(),
+                    "Internal server error")
+            ));
         }
     }
 
     @GetMapping("/get-books")
-    public ResponseEntity<?> getAllWishlistBooks(Principal currentUser) {
+    public ResponseEntity<?> getAllLibraryBooks(Principal currentUser) {
         try {
-            return ResponseEntity.ok().body(libraryService.getAllLibraryBooks(currentUser));
+            return ResponseEntity.ok().body(new SuccessResponse<>(libraryService.getAllLibraryBooks(currentUser)));
         } catch (Exception e) {
-            return new ResponseEntity<String>("Something went wrong", HttpStatus.INTERNAL_SERVER_ERROR);
+            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(new ErrorResponse(new ErrorDescription(
+                    HttpStatus.INTERNAL_SERVER_ERROR.value(),
+                    "Internal server error")
+            ));
         }
     }
 }
